feat: #30 로그인 UI 추가
- LoginActivity에 있던 로직과 뷰모델을 LoginFragment로 이동
