+++ app/src/main/java/com/km/exchangediary/ui/login/LoginViewModel.kt
@@ -6,6 +6,7 @@ import com.km.exchangediary.base.BaseViewModel
 import com.km.exchangediary.domain.usecase.LoginUseCase
 import kotlinx.coroutines.launch
 
+/* LoginFragment 에서 사용 */
 class LoginViewModel(private val loginUseCase: LoginUseCase) : BaseViewModel() {
     private val _isLoginSuccessful = MutableLiveData<Boolean>()
     val isLoginSuccessful: LiveData<Boolean> = _isLoginSuccessful
