+++ app/src/main/java/com/km/exchangediary/ui/login/adapter/OnBoardingViewPagerAdapter.kt
@@ -4,13 +4,15 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.Lifecycle
 import androidx.viewpager2.adapter.FragmentStateAdapter
+import com.km.exchangediary.ui.login.LoginFragment
 import com.km.exchangediary.ui.login.OnBoardingFragment
 
 class OnBoardingViewPagerAdapter(fragmentManager: FragmentManager, lifeCycle: Lifecycle) : FragmentStateAdapter(fragmentManager, lifeCycle) {
     private val fragmentList = arrayListOf<Fragment>(
         OnBoardingFragment(0),
         OnBoardingFragment(1),
-        OnBoardingFragment(2)
+        OnBoardingFragment(2),
+        LoginFragment()
     )
 
     override fun getItemCount(): Int = fragmentList.size
