+++ app/src/main/res/layout/activity_login.xml
@@ -30,14 +30,5 @@
             app:tabIndicator="@null"
             app:tabGravity="center"/>
 
-        <Button
-            android:id="@+id/btn_kakao_login"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="카카오 로그인"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent" />
-
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
